+++ src-core/modules/meteor/deframer.h
@@ -38,6 +38,6 @@ namespace meteor
         // Return state
         int getState();
         // Perform deframing
-        std::vector<std::array<uint8_t, CADU_SIZE>> work(uint8_t *input, size_t size);
+        std::vector<std::array<uint8_t, CADU_SIZE>> work(uint8_t *input, int size);
     };
 } // namespace meteor
